# Remove `isLeaf` flag from rtree nodes