## Description

The flag previously indicated that a node was a leaf (i.e. has no children). However, this information can just as easily be obtained on a per node-entry basis by examining whether the `child` pointer is `nil` or not.

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

- Updated release notes? (if appropriate?) N/A

## Related Issue

- N/A

## Benchmark Results

No real difference:

<details>

<summary>Click to expand</summary>

```
goos: linux
goarch: arm64
pkg: github.com/peterstace/simplefeatures/geom
                                                           │ /tmp/tmp.WtGzXnDBiu │          /tmp/tmp.S0S0fYjdZK          │
                                                           │       sec/op        │     sec/op      vs base               │
LineEnvelope/0-2                                                   2.121n ±   8%    2.145n ±   4%       ~ (p=0.151 n=15)
LineEnvelope/1-2                                                   2.131n ±   1%    2.115n ±   1%       ~ (p=0.075 n=15)
LineEnvelope/2-2                                                   2.096n ±   1%    2.097n ±   1%       ~ (p=0.659 n=15)
LineEnvelope/3-2                                                   2.103n ±   1%    2.097n ±   1%       ~ (p=0.244 n=15)
MarshalWKB/polygon/n=10-2                                          92.92n ±   1%    92.79n ±   1%       ~ (p=0.486 n=15)
MarshalWKB/polygon/n=100-2                                         235.1n ±   5%    238.2n ±   2%       ~ (p=0.361 n=15)
MarshalWKB/polygon/n=1000-2                                        1.588µ ±   7%    1.674µ ±   6%       ~ (p=0.177 n=15)
MarshalWKB/polygon/n=10000-2                                       40.80µ ±  10%    42.95µ ±  37%       ~ (p=0.486 n=15)
UnmarshalWKB/polygon/n=10-2                                        158.6n ±   1%    156.7n ±   1%  -1.20% (p=0.034 n=15)
UnmarshalWKB/polygon/n=100-2                                       300.9n ±   4%    304.1n ±   6%       ~ (p=0.532 n=15)
UnmarshalWKB/polygon/n=1000-2                                      1.863µ ±   6%    1.796µ ±   4%       ~ (p=0.123 n=15)
UnmarshalWKB/polygon/n=10000-2                                     48.58µ ±  23%    46.32µ ±  17%       ~ (p=0.935 n=15)
IntersectsLineStringWithLineString/n=10-2                          714.6n ±   1%    719.3n ±   1%       ~ (p=0.546 n=15)
IntersectsLineStringWithLineString/n=100-2                         10.51µ ±   2%    10.53µ ±   1%       ~ (p=0.653 n=15)
IntersectsLineStringWithLineString/n=1000-2                        136.4µ ±   4%    139.0µ ±   5%       ~ (p=0.436 n=15)
IntersectsLineStringWithLineString/n=10000-2                       2.078m ±  11%    2.136m ±   7%       ~ (p=0.389 n=15)
IntersectsMultiPointWithMultiPoint/n=20-2                          557.5n ±   1%    559.3n ±   1%       ~ (p=0.910 n=15)
IntersectsMultiPointWithMultiPoint/n=200-2                         7.497µ ±   1%    7.528µ ±   0%       ~ (p=0.992 n=15)
IntersectsMultiPointWithMultiPoint/n=2000-2                        71.70µ ±   1%    71.43µ ±   1%       ~ (p=0.806 n=15)
IntersectsMultiPointWithMultiPoint/n=20000-2                       780.8µ ±   1%    779.0µ ±   0%       ~ (p=0.744 n=15)
PolygonSingleRingValidation/n=10-2                                 1.233µ ±   1%    1.228µ ±   2%       ~ (p=0.466 n=15)
PolygonSingleRingValidation/n=100-2                                17.99µ ±   2%    17.83µ ±   1%       ~ (p=0.752 n=15)
PolygonSingleRingValidation/n=1000-2                               282.0µ ±   1%    283.7µ ±   1%       ~ (p=0.089 n=15)
PolygonSingleRingValidation/n=10000-2                              3.407m ±   1%    3.442m ±   1%  +1.02% (p=0.019 n=15)
PolygonMultipleRingsValidation/n=4-2                               3.500µ ±   1%    3.483µ ±   0%       ~ (p=0.602 n=15)
PolygonMultipleRingsValidation/n=36-2                              29.73µ ±   2%    29.79µ ±   2%       ~ (p=0.838 n=15)
PolygonMultipleRingsValidation/n=400-2                             385.9µ ±   3%    386.8µ ±   1%       ~ (p=0.486 n=15)
PolygonMultipleRingsValidation/n=4096-2                            4.541m ±   1%    4.574m ±   1%       ~ (p=0.148 n=15)
PolygonZigZagRingsValidation/n=10-2                                5.939µ ±   1%    5.919µ ±   1%       ~ (p=0.285 n=15)
PolygonZigZagRingsValidation/n=100-2                               68.79µ ±   2%    69.86µ ±   1%       ~ (p=0.089 n=15)
PolygonZigZagRingsValidation/n=1000-2                              981.1µ ±   2%   1001.4µ ±   2%       ~ (p=0.106 n=15)
PolygonZigZagRingsValidation/n=10000-2                             12.62m ±   2%    12.64m ±   1%       ~ (p=0.870 n=15)
PolygonAnnulusValidation/n=10-2                                    1.867µ ±   1%    1.871µ ±   1%       ~ (p=0.927 n=15)
PolygonAnnulusValidation/n=100-2                                   18.78µ ±   4%    18.87µ ±   1%       ~ (p=0.911 n=15)
PolygonAnnulusValidation/n=1000-2                                  304.3µ ±   1%    306.3µ ±   1%       ~ (p=0.389 n=15)
PolygonAnnulusValidation/n=10000-2                                 4.316m ±   2%    4.326m ±   2%       ~ (p=0.305 n=15)
MultipolygonValidation/n=1-2                                       198.5n ±   1%    197.5n ±   1%       ~ (p=0.660 n=15)
MultipolygonValidation/n=4-2                                       473.4n ±   1%    475.0n ±   1%       ~ (p=0.766 n=15)
MultipolygonValidation/n=16-2                                      2.095µ ±   1%    2.096µ ±   2%       ~ (p=0.455 n=15)
MultipolygonValidation/n=64-2                                      10.19µ ±   2%    10.14µ ±   4%       ~ (p=0.512 n=15)
MultipolygonValidation/n=256-2                                     59.42µ ±   4%    58.43µ ±   1%  -1.66% (p=0.000 n=15)
MultipolygonValidation/n=1024-2                                    339.9µ ±   3%    346.1µ ±   6%       ~ (p=0.126 n=15)
MultiPolygonTwoCircles/n=10-2                                      1.828µ ±   1%    1.825µ ±   1%       ~ (p=0.879 n=15)
MultiPolygonTwoCircles/n=100-2                                     20.14µ ±   2%    20.29µ ±   2%       ~ (p=0.595 n=15)
MultiPolygonTwoCircles/n=1000-2                                    280.5µ ±   2%    285.4µ ±   4%       ~ (p=0.187 n=15)
MultiPolygonTwoCircles/n=10000-2                                   4.734m ±   8%    4.731m ±  11%       ~ (p=0.713 n=15)
MultiPolygonMultipleTouchingPoints/n=1-2                           2.532µ ±   1%    2.529µ ±   1%       ~ (p=0.798 n=15)
MultiPolygonMultipleTouchingPoints/n=10-2                          20.62µ ±   1%    20.45µ ±   1%       ~ (p=0.559 n=15)
MultiPolygonMultipleTouchingPoints/n=100-2                         259.2µ ±   1%    259.8µ ±   3%       ~ (p=0.345 n=15)
MultiPolygonMultipleTouchingPoints/n=1000-2                        3.208m ±   2%    3.180m ±   1%       ~ (p=0.202 n=15)
WKTParsing/point-2                                                 1.124µ ±   1%    1.193µ ±   2%  +6.14% (p=0.000 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=false-2                25.09µ ±   2%    24.78µ ±   2%       ~ (p=0.148 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=true-2                 24.82µ ±   2%    25.00µ ±   1%       ~ (p=0.870 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-2               464.5µ ±   3%    463.4µ ±   1%       ~ (p=0.233 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=true-2                463.6µ ±   2%    461.7µ ±   2%       ~ (p=0.389 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-2          3.131µ ±   3%    3.123µ ±   0%       ~ (p=0.878 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-2           3.129µ ±   2%    3.122µ ±   1%       ~ (p=0.361 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-2         33.11µ ±   2%    33.38µ ±   3%       ~ (p=0.281 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-2          33.16µ ±   5%    33.22µ ±   1%       ~ (p=0.480 n=15)
MultiLineStringIsSimpleManyLineStrings/n=100-2                     23.43µ ±   2%    23.68µ ±   4%  +1.09% (p=0.045 n=15)
MultiLineStringIsSimpleManyLineStrings/n=1000-2                    306.9µ ±   3%    308.0µ ±   2%       ~ (p=0.389 n=15)
ForceCWandForceCCW/0-2                                             16.11n ±   2%    16.12n ±   1%       ~ (p=0.690 n=15)
ForceCWandForceCCW/0#01-2                                          101.3n ±   1%    101.3n ±   2%       ~ (p=0.690 n=15)
ForceCWandForceCCW/1-2                                             16.20n ± 517%    16.25n ±   1%       ~ (p=0.262 n=15)
ForceCWandForceCCW/1#01-2                                          100.8n ±  84%    102.1n ±   2%       ~ (p=0.100 n=15)
ForceCWandForceCCW/2-2                                             29.68n ±   1%    29.89n ± 421%  +0.71% (p=0.003 n=15)
ForceCWandForceCCW/2#01-2                                          155.4n ±   1%    155.0n ±  81%       ~ (p=0.546 n=15)
ForceCWandForceCCW/3-2                                             29.77n ± 420%    29.89n ±   0%       ~ (p=0.736 n=15)
ForceCWandForceCCW/3#01-2                                          155.3n ±  76%    156.8n ±   2%       ~ (p=0.104 n=15)
ForceCWandForceCCW/4-2                                             43.27n ±   0%    43.42n ±   1%  +0.35% (p=0.027 n=15)
ForceCWandForceCCW/4#01-2                                          240.2n ±   0%    240.4n ±   1%       ~ (p=0.943 n=15)
ForceCWandForceCCW/5-2                                             43.33n ±  79%    43.51n ±   1%       ~ (p=0.184 n=15)
ForceCWandForceCCW/5#01-2                                          240.0n ±   1%    240.6n ±   2%       ~ (p=0.705 n=15)
ForceCWandForceCCW/6-2                                             62.36n ±   1%    62.50n ±   1%       ~ (p=0.713 n=15)
ForceCWandForceCCW/6#01-2                                          390.5n ±   3%    387.3n ±   1%       ~ (p=0.237 n=15)
EnvelopeTransformXY-2                                              140.4n ±   1%    139.6n ±   1%       ~ (p=0.108 n=15)
WKBParse/0-2                                                       422.2n ±   1%    422.2n ±   1%       ~ (p=0.862 n=15)
WKBParse/1-2                                                       421.2n ±   2%    418.6n ±   1%       ~ (p=0.233 n=15)
WKBParse/2-2                                                       484.8n ±   1%    485.7n ±   1%       ~ (p=0.721 n=15)
WKBParse/3-2                                                       486.7n ±   1%    484.0n ±   1%       ~ (p=0.705 n=15)
WKBParse/4-2                                                       557.6n ±   2%    559.9n ±   1%       ~ (p=0.976 n=15)
WKBParse/5-2                                                       429.3n ±   1%    426.8n ±   1%       ~ (p=0.767 n=15)
WKBParse/6-2                                                       503.2n ±   7%    499.8n ±   1%       ~ (p=0.237 n=15)
WKBParse/7-2                                                       487.1n ±   3%    488.1n ±   1%       ~ (p=0.976 n=15)
WKBParse/8-2                                                       586.3n ±   1%    582.1n ±   1%       ~ (p=0.361 n=15)
WKBParse/9-2                                                       311.2n ±   3%    312.3n ±   1%       ~ (p=0.506 n=15)
WKBParse/10-2                                                      311.5n ±   2%    313.5n ±   1%       ~ (p=0.911 n=15)
WKBParse/11-2                                                      313.5n ±   1%    312.0n ±   1%       ~ (p=0.406 n=15)
WKBParse/12-2                                                      306.6n ±   1%    307.6n ±   1%       ~ (p=0.329 n=15)
WKBParse/13-2                                                      596.6n ±   1%    594.8n ±   1%       ~ (p=0.798 n=15)
WKBParse/14-2                                                      712.4n ±   2%    708.3n ±   1%       ~ (p=0.205 n=15)
WKBParse/15-2                                                      716.0n ±   1%    716.8n ±   1%       ~ (p=0.783 n=15)
WKBParse/16-2                                                      852.5n ±   1%    848.1n ±   1%       ~ (p=0.493 n=15)
WKBParse/17-2                                                      705.4n ±   1%    700.8n ±   1%       ~ (p=0.300 n=15)
WKBParse/18-2                                                      909.9n ±   2%    915.5n ±   1%       ~ (p=0.675 n=15)
WKBParse/19-2                                                      908.7n ±   2%    904.2n ±   1%       ~ (p=0.202 n=15)
WKBParse/20-2                                                      1.094µ ±   5%    1.090µ ±   0%       ~ (p=0.349 n=15)
WKBParse/21-2                                                      302.6n ±   1%    300.5n ±   1%       ~ (p=0.602 n=15)
WKBParse/22-2                                                      304.4n ±   1%    303.8n ±   1%       ~ (p=0.720 n=15)
WKBParse/23-2                                                      304.5n ±   1%    300.7n ±   2%  -1.25% (p=0.029 n=15)
WKBParse/24-2                                                      298.3n ±   1%    297.5n ±   1%       ~ (p=0.300 n=15)
WKBParse/25-2                                                      1.567µ ±   3%    1.575µ ±   1%       ~ (p=0.506 n=15)
WKBParse/26-2                                                      1.979µ ±   1%    1.984µ ±   1%       ~ (p=0.645 n=15)
WKBParse/27-2                                                      1.983µ ±   1%    1.979µ ±   0%       ~ (p=0.304 n=15)
WKBParse/28-2                                                      2.464µ ±   1%    2.458µ ±   0%       ~ (p=0.674 n=15)
WKBParse/29-2                                                      301.1n ±   2%    301.8n ±   1%       ~ (p=0.546 n=15)
WKBParse/30-2                                                      304.1n ±   1%    303.1n ±   1%       ~ (p=0.830 n=15)
WKBParse/31-2                                                      304.2n ±   1%    302.7n ±   1%       ~ (p=0.721 n=15)
WKBParse/32-2                                                      299.0n ±   1%    298.8n ±   1%       ~ (p=0.467 n=15)
WKBParse/33-2                                                      570.0n ±   1%    570.0n ±   2%       ~ (p=0.480 n=15)
WKBParse/34-2                                                      664.0n ±   1%    664.0n ±   1%       ~ (p=0.660 n=15)
WKBParse/35-2                                                      659.2n ±   1%    653.6n ±   1%       ~ (p=0.233 n=15)
WKBParse/36-2                                                      721.7n ±   1%    712.8n ±   2%       ~ (p=0.412 n=15)
WKBParse/37-2                                                      811.0n ±   1%    810.3n ±   1%       ~ (p=0.838 n=15)
WKBParse/38-2                                                      969.0n ±   2%    975.3n ±   1%       ~ (p=0.505 n=15)
WKBParse/39-2                                                      968.8n ±   1%    970.5n ±   1%       ~ (p=0.519 n=15)
WKBParse/40-2                                                      1.091µ ±   1%    1.088µ ±   1%       ~ (p=0.289 n=15)
WKBParse/41-2                                                      303.4n ±   1%    303.5n ±   1%       ~ (p=0.506 n=15)
WKBParse/42-2                                                      303.6n ±   2%    307.2n ±   1%       ~ (p=0.213 n=15)
WKBParse/43-2                                                      304.7n ±   1%    306.5n ±   1%       ~ (p=0.546 n=15)
WKBParse/44-2                                                      301.8n ±   2%    299.1n ±   1%       ~ (p=0.631 n=15)
WKBParse/45-2                                                      896.0n ±   1%    888.1n ±   2%       ~ (p=0.798 n=15)
WKBParse/46-2                                                      1.086µ ±   1%    1.080µ ±   1%       ~ (p=0.783 n=15)
WKBParse/47-2                                                      1.082µ ±   2%    1.083µ ±   1%       ~ (p=0.927 n=15)
WKBParse/48-2                                                      1.244µ ±   1%    1.238µ ±   1%       ~ (p=0.645 n=15)
WKBParse/49-2                                                      1.278µ ±   2%    1.271µ ±   1%       ~ (p=0.493 n=15)
WKBParse/50-2                                                      1.602µ ±   1%    1.616µ ±   1%       ~ (p=0.383 n=15)
WKBParse/51-2                                                      1.609µ ±   2%    1.604µ ±   1%       ~ (p=0.198 n=15)
WKBParse/52-2                                                      1.874µ ±   1%    1.876µ ±   1%       ~ (p=0.846 n=15)
WKBParse/53-2                                                      303.2n ±   2%    302.3n ±   1%       ~ (p=0.345 n=15)
WKBParse/54-2                                                      307.4n ±   1%    304.6n ±   1%       ~ (p=0.134 n=15)
WKBParse/55-2                                                      305.8n ±   1%    303.3n ±   1%       ~ (p=0.063 n=15)
WKBParse/56-2                                                      302.4n ±   1%    300.9n ±   1%       ~ (p=0.129 n=15)
WKBParse/57-2                                                      1.931µ ±   1%    1.931µ ±   1%       ~ (p=0.660 n=15)
WKBParse/58-2                                                      2.375µ ±   1%    2.368µ ±   1%       ~ (p=0.546 n=15)
WKBParse/59-2                                                      2.367µ ±   1%    2.367µ ±   0%       ~ (p=0.846 n=15)
WKBParse/60-2                                                      2.864µ ±   1%    2.855µ ±   0%       ~ (p=0.069 n=15)
WKBParse/61-2                                                      302.8n ±   1%    301.8n ±   1%       ~ (p=0.751 n=15)
WKBParse/62-2                                                      305.6n ±   1%    303.6n ±   1%       ~ (p=0.406 n=15)
WKBParse/63-2                                                      303.6n ±   1%    302.5n ±   1%       ~ (p=0.300 n=15)
WKBParse/64-2                                                      299.6n ±   1%    301.3n ±   1%       ~ (p=0.894 n=15)
WKBParse/65-2                                                      881.9n ±   1%    883.2n ±   1%       ~ (p=0.838 n=15)
WKBParse/66-2                                                      1.035µ ±   1%    1.028µ ±   1%       ~ (p=0.205 n=15)
WKBParse/67-2                                                      1.017µ ±   1%    1.018µ ±   1%       ~ (p=0.689 n=15)
WKBParse/68-2                                                      1.142µ ±   1%    1.142µ ±   1%       ~ (p=0.959 n=15)
geomean                                                            1.912µ           1.914µ         +0.08%

                                                           │ /tmp/tmp.WtGzXnDBiu │          /tmp/tmp.S0S0fYjdZK          │
                                                           │        B/op         │     B/op      vs base                 │
LineEnvelope/0-2                                                    0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=15) ¹
LineEnvelope/1-2                                                    0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=15) ¹
LineEnvelope/2-2                                                    0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=15) ¹
LineEnvelope/3-2                                                    0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=10-2                                           232.0 ± 0%       232.0 ± 0%       ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=100-2                                        1.789Ki ± 0%     1.789Ki ± 0%       ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=1000-2                                       16.04Ki ± 0%     16.04Ki ± 0%       ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=10000-2                                      160.0Ki ± 0%     160.0Ki ± 0%       ~ (p=1.000 n=15) ¹
UnmarshalWKB/polygon/n=10-2                                         284.0 ± 0%       284.0 ± 0%       ~ (p=1.000 n=15) ¹
…